+++ src/Language/Haskell/TH/Instances.hs
@@ -57,15 +57,17 @@
 -- collisions of orphans between @th-orphans@ and @th-lift-instances@.
 module Language.Haskell.TH.Instances () where
 
-import Language.Haskell.TH
+import Language.Haskell.TH hiding (newName)
 import Language.Haskell.TH.Instances.Internal
 import Language.Haskell.TH.Lift (deriveLiftMany)
 import Language.Haskell.TH.ReifyMany
-import Language.Haskell.TH.Syntax
+import Language.Haskell.TH.Syntax hiding (newName)
+import Language.Haskell.TH.Syntax.Compat (Quote(..))
 
 import Control.Monad.Reader (ReaderT(ReaderT), runReaderT)
 import Control.Monad.RWS (RWST(RWST), runRWST)
 import Control.Monad.State (StateT(StateT), runStateT)
+import qualified Control.Monad.Trans as Trans (MonadTrans(lift))
 import Control.Monad.Writer (WriterT(WriterT), runWriterT)
 import Instances.TH.Lift ()
 
@@ -120,6 +122,7 @@ import qualified Control.Monad.Fail as Fail
 #if MIN_VERSION_template_haskell(2,16,0)
 import GHC.Ptr (Ptr(Ptr))
 import GHC.ForeignPtr (newForeignPtr_)
+import Language.Haskell.TH.Syntax.Compat (liftTypedFromUntypedSplice)
 import System.IO.Unsafe (unsafePerformIO)
 #endif
 
@@ -131,10 +134,10 @@ import System.IO.Unsafe (unsafeInterleaveIO)
 
 import qualified Data.Semigroup as Semi
 
-#if MIN_VERSION_base(4,11,0)
+# if MIN_VERSION_base(4,11,0)
 import Control.Exception (throwIO, catch)
 import GHC.IO.Exception (BlockedIndefinitelyOnMVar (..), FixIOException (..))
-#endif
+# endif
 #endif
 
 #if !MIN_VERSION_template_haskell(2,11,0)
@@ -400,18 +403,26 @@ instance Applicative PprM where
 deriving instance Bounded Extension
 #endif
 
+instance Quote m => Quote (ReaderT r m) where
+    newName = Trans.lift . newName
 $(deriveQuasiTrans
     [t| forall r m. Quasi m => Proxy2 (ReaderT r m) |]
     [e| \m1 m2 -> ReaderT $ \ r -> runReaderT m1 r `qRecover` runReaderT m2 r |])
 
+instance (Quote m, Monoid w) => Quote (WriterT w m) where
+    newName = Trans.lift . newName
 $(deriveQuasiTrans
     [t| forall w m. (Quasi m, Monoid w) => Proxy2 (WriterT w m) |]
     [e| \m1 m2 -> WriterT $ runWriterT m1 `qRecover` runWriterT m2 |])
 
+instance Quote m => Quote (StateT s m) where
+    newName = Trans.lift . newName
 $(deriveQuasiTrans
     [t| forall s m. Quasi m => Proxy2 (StateT s m) |]
     [e| \m1 m2 -> StateT $ \ s -> runStateT m1 s `qRecover` runStateT m2 s |])
 
+instance (Quote m, Monoid w) => Quote (RWST r w s m) where
+    newName = Trans.lift . newName
 $(deriveQuasiTrans
     [t| forall r w s m. (Quasi m, Monoid w) => Proxy2 (RWST r w s m) |]
     [e| \m1 m2 -> RWST $ \ r s -> runRWST m1 r s `qRecover` runRWST m2 r s |])
@@ -433,7 +444,7 @@ instance Lift Bytes where
     |]
     where
       size = bytesSize bytes
-  liftTyped = unsafeTExpCoerce . lift
+  liftTyped = liftTypedFromUntypedSplice
 #endif
 
 #if !MIN_VERSION_template_haskell(2,17,0)